minor reorganization of tasks
