+++ nlp/custom_tasks/__init__.py
@@ -5,7 +5,7 @@ from .ColumbiaTransfusionTask import ColumbiaTransfusionTask
 from .TextStatsTask import *
 from .RaceFinderTask import RaceFinderTask
 from .GleasonScoreTask import GleasonScoreTask
-from .TermProximityTask import TermProximityTask
 from .pft_value_extraction import *
+from .TNMStagerTask import TNMStagerTask
