+++ nlp/tasks/__init__.py
@@ -3,7 +3,7 @@ from .MeasurementFinderTask import *
 from .ValueExtractorTask import *
 from .NERTask import *
 from .POSTaggerTask import *
-from .TNMStagerTask import *
 from .NGramTask import *
+from .TermProximityTask import TermProximityTask
 from .registered_tasks import registered_pipelines, register_pipeline_task, register_tasks, registered_collectors
