+++ nlp/tasks/registered_tasks.py
@@ -22,8 +22,8 @@ registered_pipelines = {
     "NamedEntityRecognition": NERTask,
     "NER": NERTask,
     "POSTagger": POSTaggerTask,
-    "TNMStager": TNMStagerTask,
-    "ngram": NGramTask
+    "ngram": NGramTask,
+    "TermProximityTask": TermProximityTask
 }
 
 registered_collectors = {
